Release v1.3.1

+++ tests/.lib/create-source.sh
@@ -15,22 +15,35 @@ set -o pipefail
 create_file() {
 	local src_dir="${1}"
 	local file_name="${2}"
+	local sub_dir=
+	local file_path=
 	local file_size="${3}"
 	local file_perms="${4}"
 
-	local file_path="${src_dir}/${file_name}"
+	file_name="$( printf "%q" "${file_name}" )"
+	sub_dir="$( printf "%q" "$( eval "dirname ${file_name}" )" )"
+	file_path="${src_dir}/${file_name}"
 
-	# Create directory if it doesn't exist
-	if [ ! -d "$( dirname "${file_path}" )" ]; then
-		run "mkdir -p $( dirname "${file_path}" )"
+	if ! eval "test -d ${src_dir}"; then
+		printf "No such directpry: %s\\n" "${src_dir}"
+		return 1
 	fi
-
+	# Create sub-directory if it doesn't exist
+	if [ "${sub_dir}" != "." ]; then
+		printf "# Create basedir: %s\\n" "${sub_dir}"
+		run "mkdir -p ${src_dir}/${sub_dir}"
+	fi
 	if [ "$(uname)" = "Linux" ]; then
 		run "dd if=/dev/zero of=${file_path} bs=1M count=${file_size} 2>/dev/null"
 	else
 		run "dd if=/dev/zero of=${file_path} bs=1m count=${file_size} 2>/dev/null"
 	fi
-	run "chmod ${file_perms} '${file_path}'"
+	run "chmod ${file_perms} ${file_path}"
+
+	if ! eval "test -f ${file_path}"; then
+		echo "No file created: ${file_path}"
+		return 1
+	fi
 }
 
 
@@ -40,13 +53,20 @@ create_file() {
 create_link() {
 	local src_dir="${1}"
 	local link_name="${2}"
+	local sub_dir=
 	local link_from="${3}"
-	local link_path="${src_dir}/${link_name}"
 
-	# Create directory if it doesn't exist
-	if [ ! -d "$( dirname "${link_path}" )" ]; then
-		run "mkdir -p $( dirname "${link_path}" )"
+	link_name="$( printf "%q" "${link_name}" )"
+	sub_dir="$( printf "%q" "$( eval "dirname ${link_name}" )" )"
+
+	if ! eval "test -d ${src_dir}"; then
+		printf "No such directpry: %s\\n" "${src_dir}"
+		return 1
 	fi
-
-	run "cd '${src_dir}' && ln -s '${link_from}' '${link_name}'"
+	# Create sub-directory if it doesn't exist
+	if [ "${sub_dir}" != "." ]; then
+		printf "# Create basedir: %s\\n" "${sub_dir}"
+		run "mkdir -p ${src_dir}/${sub_dir}"
+	fi
+	run "cd ${src_dir} && ln -s ${link_from} ${link_name}"
 }
